requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Data collection
requests>=2.28.0
beautifulsoup4>=4.12.0
yfinance>=0.2.0
lxml>=4.9.0

# Broker APIs
kiteconnect>=4.0.0
upstox-python>=2.0.0
smartapi-python>=1.3.0

# ML and statistical models
scikit-learn>=1.2.0
xgboost>=1.7.0
tensorflow>=2.12.0
optuna>=3.1.0
ta-lib>=0.4.0

# Database and caching
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
alembic>=1.10.0

# Dashboard and visualization
streamlit>=1.22.0
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Alerts and notifications
python-telegram-bot>=20.0
sendgrid>=6.10.0

# WebSocket for real-time data
websocket-client>=1.5.0
websockets>=11.0

# Utilities
python-dotenv>=1.0.0
pydantic>=1.10.0
joblib>=1.2.0
pytz>=2023.3
python-dateutil>=2.8.0
tqdm>=4.65.0

# Async support
aiohttp>=3.8.0
asyncio>=3.4.3
